Fixed make test linting

+++ pkg/disk/stat_bsd.go
@@ -30,12 +30,12 @@ func GetInfo(path string) (info Info, err error) {
 	if err != nil {
 		return Info{}, err
 	}
-	reservedBlocks := uint64(s.Bfree) - uint64(s.Bavail)
+	reservedBlocks := s.Bfree - s.Bavail
 	info = Info{
-		Total:  uint64(s.Bsize) * (uint64(s.Blocks) - reservedBlocks),
-		Free:   uint64(s.Bsize) * uint64(s.Bavail),
-		Files:  uint64(s.Files),
-		Ffree:  uint64(s.Ffree),
+		Total:  uint64(s.Bsize) * (s.Blocks - reservedBlocks),
+		Free:   uint64(s.Bsize) * s.Bavail,
+		Files:  s.Files,
+		Ffree:  s.Ffree,
 		FSType: getFSType(s.Fstypename[:]),
 	}
 	if info.Free > info.Total {
