+++ pkg/sys/stats_darwin.go
@@ -33,7 +33,7 @@ func getHwMemsize() (uint64, error) {
 	// removes the last byte of the result if it's 0 :/
 	totalString += "\x00"
 
-	total := uint64(binary.LittleEndian.Uint64([]byte(totalString)))
+	total := binary.LittleEndian.Uint64([]byte(totalString))
 
 	return total, nil
 }
